♻️ Extract ShiftAttendeeRow from ShiftAgendaViewer
シフト表アジェンダに埋め込まれていた出勤者行の表示ロジック（部門アイコン＋シフト種別名＋割り当て済みインストラクター名、自分は強調）を独立コンポーネント ShiftAttendeeRow へ切り出す。

- ダッシュボードの出勤状況セクション（#201/#202）と共有するための共通プリミティブ化
- 描画結果は現状と同一（純粋リファクタ・見た目の変更なし）
